fix(manifest): Properly look for manifest when working with AGP 4.1+"
Bug: BUGSEE-4927

Android Gradle Plugin (AGP) 4.1 changed the way path to the
manifest can be obtained. As the result, our plugin failed
to do it correctly which in turn led to the build failure.

This change brings the new way to obtain the path to a
manifest with multiple fallbacks for older AGP versions.

Change-Id: I3b8ba6b4c677a7878fa50f3e85c157d94f877a71