fix(manifest): Pick the right manifest to inject build UUID
Initially, I added the block which should pick the right manifest
for the Android Gradle Plugin 4.1, but it seems that it did not do
that correctly. However, other fallbacks following that block
seems to be doing the right thing. In all my tests it picked the
correct manifest and injected UUID in it

Change-Id: I85c120680318debfa9cd47d160be2156dba0d0d1